smaller scope refactor
