+++ iris-mpc/src/lib.rs
@@ -1,1 +1,3 @@
 pub mod client;
+pub mod utils;
+pub mod aws;
